Move material components into `googlesitekit-components`.
