+++ assets/js/googlesitekit/components-gm2/ProgressBar.js
