+++ webpack.config.js
@@ -135,6 +135,7 @@ const siteKitExternals = {
 	'googlesitekit-modules': [ 'googlesitekit', 'modules' ],
 	'googlesitekit-widgets': [ 'googlesitekit', 'widgets' ],
 	'@wordpress/i18n': [ 'googlesitekit', 'i18n' ],
+	'googlesitekit-components': [ 'googlesitekit', 'components' ],
 };
 
 const externals = { ...siteKitExternals };
@@ -299,6 +300,8 @@ function* webpackConfig( env, argv ) {
 			'googlesitekit-wp-dashboard':
 				'./assets/js/googlesitekit-wp-dashboard.js',
 			'googlesitekit-base': './assets/js/googlesitekit-base.js',
+			'googlesitekit-components':
+				'/assets/js/googlesitekit-components-gm2.js',
 		},
 		externals,
 		output: {
